blalalalalalala

+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: zhelm <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/22 07:03:30 by zhelm             #+#    #+#             */
-/*   Updated: 2019/05/22 09:22:36 by zhelm            ###   ########.fr       */
+/*   Updated: 2019/05/23 11:30:25 by zhelm            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 #include <string.h>
 void *ft_memset(void *b, int c, size_t len)
 {
-	int i;
+	size_t i;
 	char *s = (char *)b;
 	unsigned char f = (unsigned char)c;
 	i = 0;
